// packages/render-vue2/src/interface.ts
export type MetaDataType = keyof typeof MetaDataTypeEnum;

export type StrNumBool = '0' | '1' | 1 | 0 | boolean | undefined | null;
// ...
export interface MetaSchemaGroup {
  code: string; // 分组编码
  name: string; // 分组名称
  authCode?: string; // 权限编码
  hiddenName?: boolean; // 是否隐藏名称
  component?: string; // 组件
  mode?: typeof TABS_GROUP_MODE | typeof STEPS_GROUP_MODE;
  modeCodes?: string[];
  componentProps?: {
    // 组件属性
    [key: string]: any;
  };
  logics?: MetaLogic[]; // 页面逻辑
  hidden: boolean;
  originCode?: string; // 因不支持.命名的字段所以会把.改成-
}

export interface MetaLogic {
  effectHook: EffectHook;
  logicCode: string;
  hasChildren?: StrNumBool;
  params?: Record<any, any>;
}

export interface ValidateRules {
  validatorRule: string;
  validatorRuleValue?: any;
  triggerType?: FieldFeedbackTriggerTypes;
  message?: string;
}

export interface MetaSchemaData {
  code: string; // 字段
  originCode?: string; // 因不支持.命名的字段所以会把.改成-
  name: string; // 字段名称
  type: MetaDataType; // 字段类型
  schemaType: SchemaTypes; // 字段值类型
  group?: string; // 所属分组
  hiddenName?: StrNumBool; // 是否隐藏名称 1 是 0 否
  disabled?: StrNumBool; // 是否禁止
  required?: StrNumBool; // 是否必填
  hidden?: StrNumBool; // 是否隐藏字段
  hasSerialNo?: StrNumBool; // 是否需要显示序号
  hasSort?: StrNumBool; // 是否需要拖动排序
  description?: string; // 字段补充描述
  colSpan?: number; // 跨列
  wrap?: StrNumBool; // 是否换行
  labelCol?: number; // 名称占比宽度
  wrapperCol?: number; // 组件占比宽度
  defaultValue?: any; // 默认值
  component?: string; // 组件
  parentCode?: string; // 上一级编码
  slotName?: string;
  logics?: MetaLogic[]; // 页面逻辑
  authCode?: string; // 权限编码
  componentProps?: {
    [key: string]: any;
  }; // 组件属性
  validator?: ValidateRules[]; // 字段校验
  pageCode?: string; // 页面编码使用场景是当 schemaType 是 object 或者 array 时如果无法配置出来进行拆分配置。
  itemMetaSchema?: MetaSchema; // 使用场景是当 schemaType 是 object 或者 array 他们子项的数据集合
  agDetailMetaSchema?: MetaSchema; // 使用场景是当 schemaType ag 表格详情页
}

export interface MetaSchema {
  code: string;
  name: string;
  model: string;
  columnLayout?: number; // 几列布局
  defaultSearchColumn?: number; // 默认搜索数量
  labelCol?: number | undefined;
  wrapperCol?: number | undefined;
  remark?: string;
  group?: MetaSchemaGroup[];
  data: MetaSchemaData[];
  layoutComponent?: string;
  layoutComponentProps?: Record<string, any>;
}